frontend/requirements.txt
fastapi
uvicorn
jinja2
httpx
python-dotenv
requests
python-multipart